+++ CHANGELOG.md
@@ -23,6 +23,11 @@
 - switched in wait from webdriver TimeoutException to selene.core.exceptions.TimeoutException
   - actually no need to reuse webdriver one
   - and this might help with reporting selene failure in allure reports, let's see...
+- tried to implement something special for configuring remote driver management in shared config... 
+  - but... just left some comments for future... 
+  - it's too complicated to be implemennted in a consistent way in selene. 
+  - so far the main strategy is just to create an instance on your own 
+  - and then set it in config by config.driver = webdriver.Remote(....), KISS ;)
 
 ## 2.0.0a13 (released on 10.01.2020)
 - added temporary Collection#filter_by as deprecated